cmd: Update dev only commands
