+++ internal/cmd/root/root.go
@@ -7,7 +7,6 @@ import (
 	"os"
 
 	"github.com/gittuf/gittuf/internal/cmd/addhooks"
-	"github.com/gittuf/gittuf/internal/cmd/authorize"
 	"github.com/gittuf/gittuf/internal/cmd/clone"
 	"github.com/gittuf/gittuf/internal/cmd/dev"
 	"github.com/gittuf/gittuf/internal/cmd/policy"
@@ -49,7 +48,6 @@ func New() *cobra.Command {
 	o.AddFlags(cmd)
 
 	cmd.AddCommand(addhooks.New())
-	cmd.AddCommand(authorize.New())
 	cmd.AddCommand(clone.New())
 	cmd.AddCommand(dev.New())
 	cmd.AddCommand(trust.New())
