+++ internal/cmd/rsl/record/record.go
@@ -3,61 +3,19 @@
 package record
 
 import (
-	"fmt"
-	"os"
-
 	"github.com/gittuf/gittuf/internal/cmd/common"
-	"github.com/gittuf/gittuf/internal/dev"
 	"github.com/gittuf/gittuf/internal/repository"
 	"github.com/spf13/cobra"
 )
 
-const (
-	commitFlagName = "commit"
-)
+type options struct{}
 
-type options struct {
-	commitID       string
-	signingKeyPath string
-}
-
-func (o *options) AddFlags(cmd *cobra.Command) {
-	cmd.Flags().StringVarP(
-		&o.commitID,
-		commitFlagName,
-		"c",
-		"",
-		fmt.Sprintf("commit ID (developer mode only, set %s=1)", dev.DevModeKey),
-	)
-
-	cmd.Flags().StringVarP(
-		&o.signingKeyPath,
-		"signing-key",
-		"k",
-		"",
-		fmt.Sprintf("path to PEM encoded SSH or GPG signing key (developer mode only, set %s=1)", dev.DevModeKey),
-	)
-
-	cmd.MarkFlagsRequiredTogether("commit", "signing-key")
-}
+func (o *options) AddFlags(_ *cobra.Command) {}
 
 func (o *options) Run(_ *cobra.Command, args []string) error {
 	repo, err := repository.LoadRepository()
 	if err != nil {
 		return err
-	}
-
-	if len(o.commitID) > 0 {
-		if !dev.InDevMode() {
-			return dev.ErrNotInDevMode
-		}
-
-		signingKeyBytes, err := os.ReadFile(o.signingKeyPath)
-		if err != nil {
-			return err
-		}
-
-		return repo.RecordRSLEntryForReferenceAtCommit(args[0], o.commitID, signingKeyBytes)
 	}
 
 	return repo.RecordRSLEntryForReference(args[0], true)
@@ -69,20 +27,11 @@ func New() *cobra.Command {
 		Use:               "record",
 		Short:             "Record latest state of a Git reference in the RSL",
 		Args:              cobra.ExactArgs(1),
-		PreRunE:           checkIfSigningViable,
+		PreRunE:           common.CheckIfSigningViable,
 		RunE:              o.Run,
 		DisableAutoGenTag: true,
 	}
 	o.AddFlags(cmd)
 
 	return cmd
-}
-
-func checkIfSigningViable(cmd *cobra.Command, args []string) error {
-	commitFlag := cmd.Flag(commitFlagName)
-	if commitFlag.Value.String() != "" {
-		return nil
-	}
-
-	return common.CheckIfSigningViable(cmd, args)
 }
